fix compiler warnings (trivial) Signed-off-by: Stefan Reinauer <stepan@coresystems.de> Acked-by: Stefan Reinauer <stepan@coresystems.de>
git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3882 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/lib/memcpy.c
@@ -5,7 +5,7 @@ void *memcpy(void *vdest, const void *vsrc, size_t bytes)
 	char *dest = vdest;
 	int i;
 
-	for (i = 0; i < bytes; i++)
+	for (i = 0; i < (int)bytes; i++)
 		dest[i] = src[i];
 
 	return vdest;
